Avoid misaligned memory access
The old macro read_len and write_len may misaligned access the memory
which is report by UBSAN. This patch try to fix the problem by simply
accessing the memory with for loop, and we leave compiler unrolling
the loop for optimization.

However, I am wondering if there's a better way to optimize this. So
the macro can still be changed someday.