Emulate cpu clock and fix wrong related behivior
For the detailed change in this patch:
1. Emulate mtime at CLINT and TIME at CSR for our emulator
2. The machine timer irq should not be delegatated to
supervisor mode. (I'm not sure WHY? But qemu do this, and
we need it to load xv6 correctly.)
3. The workflow for whether the interrupt is enabled to be taken
should be fixed

Note that the implementation now could not efficient enough since
some value are repeatedly calculated. Optimization could be required
for the speed. On the other hand, clean code should be considered
for the readability.